modified:   hardware/arduino/serial/receive.ino 	modified:   hardware/i2c/ak8963/ak8963.c 	modified:   hardware/i2c/ak8963/makefile 	modified:   hardware/i2c/bmp180/bmp180.c 	modified:   hardware/i2c/bmp180/makefile 	modified:   hardware/i2c/bmp280/bmp280.c 	modified:   hardware/i2c/bmp280/makefile 	new file:   hardware/i2c/hmc5983/hmc5983.c 	new file:   hardware/i2c/hmc5983/makefile 	modified:   hardware/i2c/l3gd20/l3gd20.c 	modified:   hardware/i2c/l3gd20/makefile 	modified:   hardware/i2c/libi2c.c 	modified:   hardware/i2c/lsm303d/lsm303d.c 	modified:   hardware/i2c/lsm303d/makefile 	modified:   hardware/i2c/mpu6050/makefile 	modified:   hardware/i2c/mpu6050/mpu6050.c 	modified:   hardware/i2c/mpu9250/makefile 	modified:   hardware/i2c/mpu9250/mpu9250.c 	modified:   hardware/i2c/pca9685/esc.c 	modified:   hardware/i2c/pca9685/makefile 	modified:   hardware/i2c/pcf8591/makefile 	deleted:    hardware/i2c/pcf8591/raw.c 	new file:   hardware/serial/hc05/at.txt 	new file:   hardware/serial/hc05/rfcomm.txt 	modified:   hardware/serial/neo6m/neo6m.c 	modified:   hardware1/drone/2.sensor/1.ak8963.c 	modified:   hardware1/drone/2.sensor/2.lsm303d.c 	new file:   hardware1/drone/2.sensor/3.hmc5983.c 	new file:   hardware1/drone/2.sensor/3.neom8n.c

+++ hardware1/drone/2.sensor/2.lsm303d.c
@@ -7,12 +7,12 @@
 #include<linux/i2c-dev.h>
 
 //
-static float xmin=0;
-static float xmax=0;
-static float ymin=0;
-static float ymax=0;
-static float zmin=0;
-static float zmax=0;
+static int xmin=0;
+static int xmax=0;
+static int ymin=0;
+static int ymax=0;
+static int zmin=0;
+static int zmax=0;
 
 //
 static unsigned char reg[0x10];
